[DEB] Modernize Debian packaging for XRootD
The supported OSs are Debian 11 and Ubuntu 22.04 or later,
as stated in the top-level README.md. Based on downstream
Debian packaging, but with some important differences, such
as added support for erasure coding, running tests as part
of the build, and using a separate step to install Python
bindings, as the standard installation goes into the wrong
destination on Debian 11 (/usr/lib/python3.X/site-packages
instead of /usr/lib/python3/dist-packages).